Refactor clan experience handling and introduce energy gain events

+++ clans/src/main/resources/clans-migrations/postgres/V20260217_1__Add_clan_xp_contributions.sql
@@ -2,9 +2,9 @@
 -- Used for leaderboards and recognizing top contributors.
 CREATE TABLE IF NOT EXISTS clan_xp_contributions
 (
-    clan         BIGINT           NOT NULL REFERENCES clans (id) ON DELETE CASCADE,
-    member       VARCHAR(36)      NOT NULL,
-    contribution DOUBLE PRECISION NOT NULL DEFAULT 0,
+    clan         BIGINT      NOT NULL REFERENCES clans (id) ON DELETE CASCADE,
+    member       VARCHAR(36) NOT NULL,
+    contribution BIGINT      NOT NULL DEFAULT 0,
     PRIMARY KEY (clan, member)
 );
